+++ crates/acts/src/scheduler/tests/mod.rs
@@ -3,6 +3,7 @@ mod action;
 mod branch;
 mod message;
 mod proc;
+mod runtime;
 mod scher;
 mod state;
 mod step;